+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Onboarding - Add specific error logs for bad configurations.
+
 ## RELEASE 0.0.24 - 2017-07-11
 ### Added
 - Search - Users can search on the hasMany associated data of a specific record.